+++ tests/Integration/Framework/Bootstrap/ReadOnlyAppRoot/Config.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Integration\Framework\Bootstrap\ReadOnlyAppRoot;
+
+use Gacela\Framework\AbstractConfig;
+
+final class Config extends AbstractConfig
+{
+}