+++ tests/Integration/Framework/Bootstrap/ReadOnlyAppRoot/Factory.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Integration\Framework\Bootstrap\ReadOnlyAppRoot;
+
+use Gacela\Framework\AbstractFactory;
+
+/**
+ * @extends AbstractFactory<Config>
+ */
+final class Factory extends AbstractFactory
+{
+    public function createGreetingService(): GreetingService
+    {
+        return new GreetingService();
+    }
+}